requirements.txt
astroid==2.2.5
Click==7.0
Flask==1.1.1
isort==4.3.21
itsdangerous==1.1.0
JayDeBeApi==1.1.1
Jinja2==2.10.1
JPype1==0.6.3
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.3.1
six==1.12.0
typed-ast==1.4.0
Werkzeug==0.15.4
wrapt==1.11.2
